Moving similar alignment helper functions to utils file.
