+++ vsg/rules/align_tokens_in_region_between_tokens.py
@@ -94,7 +94,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
                    iColumn += alignment_utils.update_column_width(self, oToken)
 
                if isinstance(oToken, token.generic_clause.semicolon) and self.separate_generic_port_alignment:
-                   add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+                   alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
                    for iKey in list(dAnalysis.keys()):
                        if dAnalysis[iKey]['adjust'] != 0:
                            oLineTokens = oFile.get_tokens_from_line(iKey)
@@ -106,7 +106,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
                    dAnalysis = {}
 
                if isinstance(oToken, token.generic_map_aspect.close_parenthesis) and self.separate_generic_port_alignment:
-                   add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+                   alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
                    for iKey in list(dAnalysis.keys()):
                        if dAnalysis[iKey]['adjust'] != 0:
                            oLineTokens = oFile.get_tokens_from_line(iKey)
@@ -118,7 +118,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
                    dAnalysis = {}
 
                if generate_statement_detected(self, oToken):
-                   add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+                   alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
                    for iKey in list(dAnalysis.keys()):
                        if dAnalysis[iKey]['adjust'] != 0:
                            oLineTokens = oFile.get_tokens_from_line(iKey)
@@ -137,7 +137,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
                    if self.comment_line_ends_group:
                        if utils.are_next_consecutive_token_types([parser.whitespace, parser.comment], iIndex + 1, lTokens) or \
                           utils.are_next_consecutive_token_types([parser.comment], iIndex + 1, lTokens):
-                           add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+                           alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
                            for iKey in list(dAnalysis.keys()):
                                if dAnalysis[iKey]['adjust'] != 0:
                                    oLineTokens = oFile.get_tokens_from_line(iKey)
@@ -150,7 +150,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
 
                    if self.blank_line_ends_group:
                        if utils.are_next_consecutive_token_types([parser.blank_line], iIndex + 1, lTokens):
-                           add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+                           alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
 
                            for iKey in list(dAnalysis.keys()):
                                if dAnalysis[iKey]['adjust'] != 0:
@@ -162,7 +162,7 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
 
                            dAnalysis = {}
 
-            add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
+            alignment_utils.add_adjustments_to_dAnalysis(dAnalysis, self.compact_alignment)
 
             for iKey in list(dAnalysis.keys()):
                 if dAnalysis[iKey]['adjust'] != 0:
@@ -186,26 +186,6 @@ class align_tokens_in_region_between_tokens(alignment.Rule):
         else:
             rules_utils.insert_whitespace(lTokens, iTokenIndex, dAction['adjust'])
         oViolation.set_tokens(lTokens)
-
-
-def add_adjustments_to_dAnalysis(dAnalysis, compact_alignment):
-    iMaxLeftColumn = 0
-    iMinLeftColumn = 9999999999999999
-    iMaxTokenColumn = 0
-    iMinTokenColumn = 9999999999999999
-
-    for iKey in list(dAnalysis.keys()):
-        iMaxLeftColumn = max(iMaxLeftColumn, dAnalysis[iKey]['left_column'])
-        iMinLeftColumn = min(iMinLeftColumn, dAnalysis[iKey]['left_column'])
-        iMaxTokenColumn = max(iMaxTokenColumn, dAnalysis[iKey]['token_column'])
-        iMinTokenColumn = min(iMinTokenColumn, dAnalysis[iKey]['token_column'])
-
-    if compact_alignment:
-        for iKey in list(dAnalysis.keys()):
-            dAnalysis[iKey]['adjust'] = iMaxLeftColumn - dAnalysis[iKey]['token_column'] + 1
-    else:
-        for iKey in list(dAnalysis.keys()):
-            dAnalysis[iKey]['adjust'] = iMaxTokenColumn - dAnalysis[iKey]['token_column']
 
 
 def generate_statement_detected(self, oToken):
